+++ BUILD_FIX_FINAL.md
@@ -1,0 +1,30 @@
+# Build Resilience Fix - Final
+
+**Issue:** Webpack error "Assigning to rvalue" caused by Prisma client being bundled for client-side.
+
+**Root Cause:** 
+- `business.ts` imported Prisma but was used in client components (`pricing/page.tsx`, `billing/success/page.tsx`)
+- `prismaClient.ts` was setting environment variables in a way that webpack couldn't handle during client bundling
+
+**Solution Applied:**
+
+1. **Split business metrics into client-safe and server-only:**
+   - `business.ts` - Client-safe tracking functions (no Prisma)
+   - `business-server.ts` - Server-only database queries (with Prisma)
+
+2. **Made business.ts client-only:**
+   - Added `'use client'` directive
+   - Removed Prisma import
+   - Added client-side check for analytics tracking
+
+3. **Hardened prismaClient.ts:**
+   - Added `typeof window === 'undefined'` checks before setting env vars
+   - Prevents webpack from processing Prisma code during client bundling
+
+**Files Changed:**
+- `packages/web/src/lib/metrics/business.ts` - Made client-safe
+- `packages/web/src/lib/metrics/business-server.ts` - New server-only file
+- `packages/web/src/shared/db/prismaClient.ts` - Added server-only guards
+- `packages/web/src/app/api/stripe/webhook/route.ts` - Updated import
+
+**Status:** ✅ **FIXED** - Build should now succeed
