+++ packages/web/src/shared/db/prismaClient.ts
@@ -15,14 +15,20 @@
 
 // CRITICAL: Set environment variables BEFORE importing PrismaClient
 // Prisma 7 determines engine type at import time, so we must set these first
-if (typeof process !== 'undefined' && process.env) {
+// Only execute in server context to prevent webpack from bundling Prisma for client
+if (typeof window === 'undefined' && typeof process !== 'undefined' && process.env) {
   // Force binary engine - this must be set before PrismaClient is imported
   // This is critical for Prisma 7 to use binary engine instead of client engine
-  process.env.PRISMA_CLIENT_ENGINE_TYPE = 'binary';
+  // Only set in server context to avoid webpack bundling issues
+  if (typeof window === 'undefined') {
+    process.env.PRISMA_CLIENT_ENGINE_TYPE = 'binary';
+  }
   
   // Ensure Node.js runtime is detected (not edge)
   // Prisma 7 uses client engine in edge/serverless environments
-  if (!process.env.NEXT_RUNTIME) {
+  // Only set during server-side execution to avoid webpack bundling issues
+  if (typeof window === 'undefined' && !process.env.NEXT_RUNTIME) {
+    // Safe to set in server context only
     process.env.NEXT_RUNTIME = 'nodejs';
   }
 
@@ -34,21 +40,24 @@ if (typeof process !== 'undefined' && process.env) {
   // Check if we're in a build context (Next.js build phase or Vercel build)
   // During Next.js build, when collecting page data, DATABASE_URL might not be set
   // but Prisma needs it to detect binary engine type
-  const isBuildPhase = 
-    process.env.NEXT_PHASE === 'phase-production-build' ||
-    (process.env.NODE_ENV === 'production' && process.env.VERCEL === '1') ||
-    (process.env.NODE_ENV === 'production' && !process.env.DATABASE_URL) ||
-    process.env.VERCEL === '1';
-  
-  if (!process.env.DATABASE_URL && isBuildPhase) {
-    // During build phase, set a dummy DATABASE_URL to help Prisma detect binary engine
-    // This is safe because we're not actually connecting during build
-    // Prisma will not actually connect during build - it only needs the URL for engine detection
-    process.env.DATABASE_URL = 'postgresql://dummy:dummy@localhost:5432/dummy?schema=public';
+  // Only execute in server context to avoid webpack bundling issues
+  if (typeof window === 'undefined') {
+    const isBuildPhase = 
+      process.env.NEXT_PHASE === 'phase-production-build' ||
+      (process.env.NODE_ENV === 'production' && process.env.VERCEL === '1') ||
+      (process.env.NODE_ENV === 'production' && !process.env.DATABASE_URL) ||
+      process.env.VERCEL === '1';
+    
+    if (!process.env.DATABASE_URL && isBuildPhase) {
+      // During build phase, set a dummy DATABASE_URL to help Prisma detect binary engine
+      // This is safe because we're not actually connecting during build
+      // Prisma will not actually connect during build - it only needs the URL for engine detection
+      process.env.DATABASE_URL = 'postgresql://dummy:dummy@localhost:5432/dummy?schema=public';
+    }
+    
+    // Store isBuildPhase for use after import
+    (globalThis as any).__PRISMA_BUILD_PHASE__ = isBuildPhase;
   }
-  
-  // Store isBuildPhase for use after import
-  (globalThis as any).__PRISMA_BUILD_PHASE__ = isBuildPhase;
 }
 
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -71,6 +80,11 @@ const isBuildPhase = (globalThis as any).__PRISMA_BUILD_PHASE__ ?? false;
 // it requires either adapter or accelerateUrl. We provide a dummy accelerateUrl
 // during build only to satisfy the constructor requirement.
 // This is safe because Prisma won't actually connect during build when collecting page data.
+// Only access isBuildPhase in server context
+const isBuildPhase = typeof window === 'undefined' 
+  ? ((globalThis as any).__PRISMA_BUILD_PHASE__ ?? false)
+  : false;
+
 const prismaConfig: ConstructorParameters<typeof PrismaClient>[0] = {
   log: process.env.NODE_ENV === 'development' ? ['query', 'error', 'warn'] : ['error'],
   // During Vercel build, provide accelerateUrl to satisfy Prisma Client constructor
